+++ src/main/java/swp391/aistudyhub/dto/projection/SystemConfigResponse.java
@@ -3,9 +3,9 @@ package swp391.aistudyhub.dto.projection;
 public interface SystemConfigResponse {
     Integer getMaxDailyChatTokens();
 
-    Long getTotalStorageQuotaGb();
+    Double getTotalStorageQuotaGb();
 
-    Long getMaxFileSizeMb();
+    Double getMaxFileSizeMb();
 
     String getAllowedFileTypes();
