+++ src/main/java/swp391/aistudyhub/entity/SystemConfig.java
@@ -22,10 +22,10 @@ public class SystemConfig {
     private Integer maxDailyChatTokens;
 
     @Column(name = "total_storage_quota_gb", nullable = false)
-    private Long totalStorageQuotaGb;
+    private Double totalStorageQuotaGb;
 
     @Column(name = "max_file_size_mb", nullable = false)
-    private Long maxFileSizeMb;
+    private Double maxFileSizeMb;
 
     @Column(name = "allowed_file_types", nullable = false)
     private String allowedFileTypes;
